+++ src/squlearn/qnn/lowlevel_qnn/__init__.py
@@ -1,0 +1,3 @@
+from .lowlevel_qnn import LowLevelQNN
+
+__all__ = ["LowLevelQNN"]
